requirements.txt
aiapy==0.7.4
aioftp==0.22.3
aiohttp==3.9.5
aiosignal==1.3.1
asciitree==0.3.3
asdf==3.2.0
asdf-astropy==0.6.1
asdf_coordinates_schemas==0.3.0
asdf_standard==1.1.1
asdf_transform_schemas==0.5.0
asdf_wcs_schemas==0.4.0
astropy==6.0.1
astropy-iers-data==0.2024.4.29.0.28.48
astropy_healpix==1.0.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cdflib==1.2.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.2.1
cycler==0.12.1
dask==2024.4.2
debugpy==1.8.20
decorator==5.1.1
drms==0.6.4
eispac==0.95.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.19
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
Functions @ file:///home/biswanath/python/Functions/dist/Functions-1.4.8-py3-none-any.whl
Glymur==0.13.1
greenlet==3.0.3
gwcs==0.21.0
h11==0.14.0
h5netcdf==1.3.0
h5py==3.11.0
halo==0.0.31
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
ipykernel==7.2.0
ipython==8.24.0
iris-obj @ file:///home/biswanath/python/iris_obj/dist/iris_obj-0.2.0-py3-none-any.whl
isodate==0.6.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.5
lazy_loader==0.4
locket==1.0.0
log-symbols==0.0.14
lxml==5.2.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpl-animators==1.1.1
multidict==6.0.5
ndcube==2.2.2
nest-asyncio==1.6.0
networkx==3.3
numcodecs==0.12.1
numpy==1.26.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parfive==2.1.0
parso==0.8.4
partd==1.4.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt_toolkit==3.0.45
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyerfa==2.0.1.4
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==27.1.0
reproject==0.13.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
scikit-image==0.23.2
scipy==1.13.0
selenium==4.23.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spinners==0.0.24
SQLAlchemy==2.0.29
stack-data==0.6.3
sunkit_image==0.6.1
sunpy==6.0.5
termcolor==2.4.0
tifffile==2024.4.24
tomli==2.0.1
toolz==0.12.1
tornado==6.5.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
zarr==2.17.2
zeep==4.2.1
zipp==3.18.1
